Корректировка модулей
